move lookup definition to cpp file
